# openapi.yaml
openapi: 3.0.1
info:
  title: Shadowy Shitcoin Advisor
  description: A plugin to track your shitcoin portfolio!
  version: 'v1'
servers:
  - url: http://localhost:5003
paths:
  /ethereum/shitcoins/{username}:
    get:
      operationId: getCoins
      summary: Get the list of coins.
      parameters:
      - in: path
        name: username
        schema:
            type: string
        required: true
        description: The name of the user.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getCoinResponse'
    post:
      operationId: addCoin
      summary: Add a coin to the list.
      parameters:
      - in: path
        name: username
        schema:
            type: string
        required: true
        description: The name of the user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/addCoinRequest'
      responses:
        "200":
          description: OK
    delete:
      operationId: deleteCoin
      summary: Delete a coin from the list.
      parameters:
      - in: path
        name: username
        schema:
            type: string
        required: true
        description: The name of the user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/removeCoinRequest'
      responses:
        "200":
          description: OK
  
components:
  schemas:
    getCoinResponse:
      type: object
      properties:
        todos:
          type: array
          items:
            type: string
          description: The list of todos.
    addCoinRequest:
      type: object
      required:
      - coin
      properties:
        coin:
          type: string
          description: The shitcoin to add to the list.
          required: true
    removeCoinRequest:
      type: object
      required:
      - coin
      properties:
        coin:
          type: string
          description: The shitcoin to remove from the list.
          required: true